@@ -18,12 +18,25 @@
   <url>http://github.com/square/javapoet/</url>
 
   <properties>
+    <maven.compiler.parameters>true</maven.compiler.parameters>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+    <maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
+    <maven-checkstyle-plugin.version>3.3.0</maven-checkstyle-plugin.version>
+    <maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
 
-    <java.version>1.8</java.version>
-    <junit.version>4.13.2</junit.version>
+    <assertj-core.version>3.24.2</assertj-core.version>
+    <junit.version>5.10.0</junit.version>
     <truth.version>1.1.5</truth.version>
     <compile-testing.version>0.21.0</compile-testing.version>
+    <mockito-core.version>5.5.0</mockito-core.version>
+    <ecj.version>4.6.1</ecj.version>
+    <jimfs.version>1.3.0</jimfs.version>
+    <checkstyle.version>10.12.3</checkstyle.version>
+    <plexus-compiler-javac-errorprone.version>2.13.0</plexus-compiler-javac-errorprone.version>
+    <error_prone_core.version>2.21.1</error_prone_core.version>
   </properties>
 
   <scm>
