@@ -15,203 +15,223 @@
  */
 package com.squareup.javapoet;
 
+import org.junit.jupiter.api.Test;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 import java.util.List;
 import java.util.Map;
-import org.junit.Test;
 
-import static com.google.common.truth.Truth.assertThat;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotEquals;
-import static org.junit.Assert.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class AnnotatedTypeNameTest {
+class AnnotatedTypeNameTest {
 
-  private final static String NN = NeverNull.class.getCanonicalName();
-  private final AnnotationSpec NEVER_NULL = AnnotationSpec.builder(NeverNull.class).build();
-  private final static String TUA = TypeUseAnnotation.class.getCanonicalName();
-  private final AnnotationSpec TYPE_USE_ANNOTATION =
-      AnnotationSpec.builder(TypeUseAnnotation.class).build();
+    private final static String NN = NeverNull.class.getCanonicalName();
+    private final AnnotationSpec NEVER_NULL = AnnotationSpec.builder(NeverNull.class).build();
+    private final static String TUA = TypeUseAnnotation.class.getCanonicalName();
+    private final AnnotationSpec TYPE_USE_ANNOTATION =
+            AnnotationSpec.builder(TypeUseAnnotation.class).build();
 
-  @Target(ElementType.TYPE_USE)
-  public @interface NeverNull {}
+    @Target(ElementType.TYPE_USE)
+    public @interface NeverNull {
+    }
 
-  @Target(ElementType.TYPE_USE)
-  public @interface TypeUseAnnotation {}
+    @Target(ElementType.TYPE_USE)
+    public @interface TypeUseAnnotation {
+    }
 
 
-  @Test(expected=NullPointerException.class) public void nullAnnotationArray() {
-    TypeName.BOOLEAN.annotated((AnnotationSpec[]) null);
-  }
+    @Test
+    void nullAnnotationArray() {
+        assertThrows(NullPointerException.class, () -> TypeName.BOOLEAN.annotated((AnnotationSpec[]) null));
+    }
 
-  @Test(expected=NullPointerException.class) public void nullAnnotationList() {
-    TypeName.DOUBLE.annotated((List<AnnotationSpec>) null);
-  }
+    @Test
+    void nullAnnotationList() {
+        assertThrows(NullPointerException.class, () -> TypeName.DOUBLE.annotated((List<AnnotationSpec>) null));
+    }
 
-  @Test public void annotated() {
-    TypeName simpleString = TypeName.get(String.class);
-    assertFalse(simpleString.isAnnotated());
-    assertEquals(simpleString, TypeName.get(String.class));
+    @Test
+    void annotated() {
+        TypeName simpleString = TypeName.get(String.class);
+        assertThat(simpleString.isAnnotated()).isFalse();
+        assertThat(TypeName.get(String.class)).isEqualTo(simpleString);
 
-    TypeName annotated = simpleString.annotated(NEVER_NULL);
-    assertTrue(annotated.isAnnotated());
-    assertEquals(annotated, annotated.annotated());
-  }
+        TypeName annotated = simpleString.annotated(NEVER_NULL);
+        assertThat(annotated.isAnnotated()).isTrue();
+        assertThat(annotated).isEqualTo(annotated.annotated());
+    }
 
-  @Test public void annotatedType() {
-    TypeName type = TypeName.get(String.class);
-    TypeName actual = type.annotated(TYPE_USE_ANNOTATION);
-    assertThat(actual.toString()).isEqualTo("java.lang. @" + TUA + " String");
-  }
+    @Test
+    void annotatedType() {
+        TypeName type = TypeName.get(String.class);
+        TypeName actual = type.annotated(TYPE_USE_ANNOTATION);
+        assertThat(actual.toString()).hasToString("java.lang. @" + TUA + " String");
+    }
 
-  @Test public void annotatedTwice() {
-    TypeName type = TypeName.get(String.class);
-    TypeName actual =
-        type.annotated(NEVER_NULL)
-            .annotated(TYPE_USE_ANNOTATION);
-    assertThat(actual.toString())
-        .isEqualTo("java.lang. @" + NN + " @" + TUA + " String");
-  }
+    @Test
+    void annotatedTwice() {
+        TypeName type = TypeName.get(String.class);
+        TypeName actual =
+                type.annotated(NEVER_NULL)
+                        .annotated(TYPE_USE_ANNOTATION);
+        assertThat(actual)
+                .hasToString("java.lang. @" + NN + " @" + TUA + " String");
+    }
 
-  @Test public void annotatedParameterizedType() {
-    TypeName type = ParameterizedTypeName.get(List.class, String.class);
-    TypeName actual = type.annotated(TYPE_USE_ANNOTATION);
-    assertThat(actual.toString()).isEqualTo("java.util. @" + TUA + " List<java.lang.String>");
-  }
+    @Test
+    void annotatedParameterizedType() {
+        TypeName type = ParameterizedTypeName.get(List.class, String.class);
+        TypeName actual = type.annotated(TYPE_USE_ANNOTATION);
+        assertThat(actual).hasToString("java.util. @" + TUA + " List<java.lang.String>");
+    }
 
-  @Test public void annotatedArgumentOfParameterizedType() {
-    TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
-    TypeName actual = ParameterizedTypeName.get(ClassName.get(List.class), type);
-    assertThat(actual.toString()).isEqualTo("java.util.List<java.lang. @" + TUA + " String>");
-  }
+    @Test
+    void annotatedArgumentOfParameterizedType() {
+        TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
+        TypeName actual = ParameterizedTypeName.get(ClassName.get(List.class), type);
+        assertThat(actual).hasToString("java.util.List<java.lang. @" + TUA + " String>");
+    }
 
-  @Test public void annotatedWildcardTypeNameWithSuper() {
-    TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
-    TypeName actual = WildcardTypeName.supertypeOf(type);
-    assertThat(actual.toString()).isEqualTo("? super java.lang. @" + TUA + " String");
-  }
+    @Test
+    void annotatedWildcardTypeNameWithSuper() {
+        TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
+        TypeName actual = WildcardTypeName.supertypeOf(type);
+        assertThat(actual).hasToString("? super java.lang. @" + TUA + " String");
+    }
 
-  @Test public void annotatedWildcardTypeNameWithExtends() {
-    TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
-    TypeName actual = WildcardTypeName.subtypeOf(type);
-    assertThat(actual.toString()).isEqualTo("? extends java.lang. @" + TUA + " String");
-  }
+    @Test
+    void annotatedWildcardTypeNameWithExtends() {
+        TypeName type = TypeName.get(String.class).annotated(TYPE_USE_ANNOTATION);
+        TypeName actual = WildcardTypeName.subtypeOf(type);
+        assertThat(actual).hasToString("? extends java.lang. @" + TUA + " String");
+    }
 
-  @Test public void annotatedEquivalence() {
-    annotatedEquivalence(TypeName.VOID);
-    annotatedEquivalence(ArrayTypeName.get(Object[].class));
-    annotatedEquivalence(ClassName.get(Object.class));
-    annotatedEquivalence(ParameterizedTypeName.get(List.class, Object.class));
-    annotatedEquivalence(TypeVariableName.get(Object.class));
-    annotatedEquivalence(WildcardTypeName.get(Object.class));
-  }
+    @Test
+    void annotatedEquivalence() {
+        annotatedEquivalence(TypeName.VOID);
+        annotatedEquivalence(ArrayTypeName.get(Object[].class));
+        annotatedEquivalence(ClassName.get(Object.class));
+        annotatedEquivalence(ParameterizedTypeName.get(List.class, Object.class));
+        annotatedEquivalence(TypeVariableName.get(Object.class));
+        annotatedEquivalence(WildcardTypeName.get(Object.class));
+    }
 
-  private void annotatedEquivalence(TypeName type) {
-    assertFalse(type.isAnnotated());
-    assertEquals(type, type);
-    assertEquals(type.annotated(TYPE_USE_ANNOTATION), type.annotated(TYPE_USE_ANNOTATION));
-    assertNotEquals(type, type.annotated(TYPE_USE_ANNOTATION));
-    assertEquals(type.hashCode(), type.hashCode());
-    assertEquals(type.annotated(TYPE_USE_ANNOTATION).hashCode(),
-        type.annotated(TYPE_USE_ANNOTATION).hashCode());
-    assertNotEquals(type.hashCode(), type.annotated(TYPE_USE_ANNOTATION).hashCode());
-  }
+    private void annotatedEquivalence(TypeName type) {
+        assertThat(type.isAnnotated()).isFalse();
+        assertThat(type).isEqualTo(type);
+        assertThat(type.annotated(TYPE_USE_ANNOTATION)).isEqualTo(type.annotated(TYPE_USE_ANNOTATION));
+        assertThat(type).isNotEqualTo(type.annotated(TYPE_USE_ANNOTATION));
+        assertThat(type).hasSameHashCodeAs(type);
+        assertThat(type.annotated(TYPE_USE_ANNOTATION).hashCode()).isEqualTo(
+                type.annotated(TYPE_USE_ANNOTATION).hashCode());
+        assertThat(type.hashCode()).isNotEqualTo(type.annotated(TYPE_USE_ANNOTATION).hashCode());
+    }
 
-  // https://github.com/square/javapoet/issues/431
-  @Test public void annotatedNestedType() {
-    TypeName type = TypeName.get(Map.Entry.class).annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.toString()).isEqualTo("java.util.Map. @" + TUA + " Entry");
-  }
+    // https://github.com/square/javapoet/issues/431
+    @Test
+    void annotatedNestedType() {
+        TypeName type = TypeName.get(Map.Entry.class).annotated(TYPE_USE_ANNOTATION);
+        assertThat(type).hasToString("java.util.Map. @" + TUA + " Entry");
+    }
 
-  @Test public void annotatedEnclosingAndNestedType() {
-    TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
-        .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.toString()).isEqualTo("java.util. @" + TUA + " Map. @" + TUA + " Entry");
-  }
+    @Test
+    void annotatedEnclosingAndNestedType() {
+        TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
+                .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
+        assertThat(type).hasToString("java.util. @" + TUA + " Map. @" + TUA + " Entry");
+    }
 
-  // https://github.com/square/javapoet/issues/431
-  @Test public void annotatedNestedParameterizedType() {
-    TypeName type = ParameterizedTypeName.get(Map.Entry.class, Byte.class, Byte.class)
-        .annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.toString())
-        .isEqualTo("java.util.Map. @" + TUA + " Entry<java.lang.Byte, java.lang.Byte>");
-  }
+    // https://github.com/square/javapoet/issues/431
+    @Test
+    void annotatedNestedParameterizedType() {
+        TypeName type = ParameterizedTypeName.get(Map.Entry.class, Byte.class, Byte.class)
+                .annotated(TYPE_USE_ANNOTATION);
+        assertThat(type)
+                .hasToString("java.util.Map. @" + TUA + " Entry<java.lang.Byte, java.lang.Byte>");
+    }
 
-  @Test public void withoutAnnotationsOnAnnotatedEnclosingAndNestedType() {
-    TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
-        .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.isAnnotated()).isTrue();
-    assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
-  }
+    @Test
+    void withoutAnnotationsOnAnnotatedEnclosingAndNestedType() {
+        TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
+                .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
+        assertThat(type.isAnnotated()).isTrue();
+        assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
+    }
 
-  @Test public void withoutAnnotationsOnAnnotatedEnclosingType() {
-    TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
-        .nestedClass("Entry");
-    assertThat(type.isAnnotated()).isTrue();
-    assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
-  }
+    @Test
+    void withoutAnnotationsOnAnnotatedEnclosingType() {
+        TypeName type = ((ClassName) TypeName.get(Map.class).annotated(TYPE_USE_ANNOTATION))
+                .nestedClass("Entry");
+        assertThat(type.isAnnotated()).isTrue();
+        assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
+    }
 
-  @Test public void withoutAnnotationsOnAnnotatedNestedType() {
-    TypeName type = ((ClassName) TypeName.get(Map.class))
-        .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.isAnnotated()).isTrue();
-    assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
-  }
+    @Test
+    void withoutAnnotationsOnAnnotatedNestedType() {
+        TypeName type = ((ClassName) TypeName.get(Map.class))
+                .nestedClass("Entry").annotated(TYPE_USE_ANNOTATION);
+        assertThat(type.isAnnotated()).isTrue();
+        assertThat(type.withoutAnnotations()).isEqualTo(TypeName.get(Map.Entry.class));
+    }
 
-  // https://github.com/square/javapoet/issues/614
-   @Test public void annotatedArrayType() {
-    TypeName type = ArrayTypeName.of(ClassName.get(Object.class)).annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.toString()).isEqualTo("java.lang.Object @" + TUA + " []");
-  }
+    // https://github.com/square/javapoet/issues/614
+    @Test
+    void annotatedArrayType() {
+        TypeName type = ArrayTypeName.of(ClassName.get(Object.class)).annotated(TYPE_USE_ANNOTATION);
+        assertThat(type).hasToString("java.lang.Object @" + TUA + " []");
+    }
 
-  @Test public void annotatedArrayElementType() {
-    TypeName type = ArrayTypeName.of(ClassName.get(Object.class).annotated(TYPE_USE_ANNOTATION));
-    assertThat(type.toString()).isEqualTo("java.lang. @" + TUA + " Object[]");
-  }
+    @Test
+    void annotatedArrayElementType() {
+        TypeName type = ArrayTypeName.of(ClassName.get(Object.class).annotated(TYPE_USE_ANNOTATION));
+        assertThat(type).hasToString("java.lang. @" + TUA + " Object[]");
+    }
 
-  // https://github.com/square/javapoet/issues/614
-  @Test public void annotatedOuterMultidimensionalArrayType() {
-    TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class)))
-        .annotated(TYPE_USE_ANNOTATION);
-    assertThat(type.toString()).isEqualTo("java.lang.Object @" + TUA + " [][]");
-  }
+    // https://github.com/square/javapoet/issues/614
+    @Test
+    void annotatedOuterMultidimensionalArrayType() {
+        TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class)))
+                .annotated(TYPE_USE_ANNOTATION);
+        assertThat(type).hasToString("java.lang.Object @" + TUA + " [][]");
+    }
 
-  // https://github.com/square/javapoet/issues/614
-  @Test public void annotatedInnerMultidimensionalArrayType() {
-    TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class))
-        .annotated(TYPE_USE_ANNOTATION));
-    assertThat(type.toString()).isEqualTo("java.lang.Object[] @" + TUA + " []");
-  }
+    // https://github.com/square/javapoet/issues/614
+    @Test
+    void annotatedInnerMultidimensionalArrayType() {
+        TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class))
+                .annotated(TYPE_USE_ANNOTATION));
+        assertThat(type).hasToString("java.lang.Object[] @" + TUA + " []");
+    }
 
-  // https://github.com/square/javapoet/issues/614
-  @Test public void annotatedArrayTypeVarargsParameter() {
-    TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class)))
-        .annotated(TYPE_USE_ANNOTATION);
-    MethodSpec varargsMethod = MethodSpec.methodBuilder("m")
-        .addParameter(
-            ParameterSpec.builder(type, "p")
-                .build())
-        .varargs()
-        .build();
-    assertThat(varargsMethod.toString()).isEqualTo(""
-        + "void m(java.lang.Object @" + TUA + " []... p) {\n"
-        + "}\n");
-  }
+    // https://github.com/square/javapoet/issues/614
+    @Test
+    void annotatedArrayTypeVarargsParameter() {
+        TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class)))
+                .annotated(TYPE_USE_ANNOTATION);
+        MethodSpec varargsMethod = MethodSpec.methodBuilder("m")
+                .addParameter(
+                        ParameterSpec.builder(type, "p")
+                                .build())
+                .varargs()
+                .build();
+        assertThat(varargsMethod).hasToString("void m(java.lang.Object @" + TUA + " []... p) {\n"
+                + "}\n");
+    }
 
-  // https://github.com/square/javapoet/issues/614
-  @Test public void annotatedArrayTypeInVarargsParameter() {
-    TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class))
-        .annotated(TYPE_USE_ANNOTATION));
-    MethodSpec varargsMethod = MethodSpec.methodBuilder("m")
-        .addParameter(
-            ParameterSpec.builder(type, "p")
-                .build())
-        .varargs()
-        .build();
-    assertThat(varargsMethod.toString()).isEqualTo(""
-        + "void m(java.lang.Object[] @" + TUA + " ... p) {\n"
-        + "}\n");
-  }
+    // https://github.com/square/javapoet/issues/614
+    @Test
+    void annotatedArrayTypeInVarargsParameter() {
+        TypeName type = ArrayTypeName.of(ArrayTypeName.of(ClassName.get(Object.class))
+                .annotated(TYPE_USE_ANNOTATION));
+        MethodSpec varargsMethod = MethodSpec.methodBuilder("m")
+                .addParameter(
+                        ParameterSpec.builder(type, "p")
+                                .build())
+                .varargs()
+                .build();
+        assertThat(varargsMethod).hasToString("void m(java.lang.Object[] @" + TUA + " ... p) {\n"
+                + "}\n");
+    }
 }
