@@ -1,19 +1,19 @@
 package com.squareup.javapoet;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 
-import static com.google.common.truth.Truth.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class CodeWriterTest {
+class CodeWriterTest {
 
     @Test
-    public void emptyLineInJavaDocDosEndings() throws IOException {
+    void emptyLineInJavaDocDosEndings() throws IOException {
         CodeBlock javadocCodeBlock = CodeBlock.of("A\r\n\r\nB\r\n");
         StringBuilder out = new StringBuilder();
         new CodeWriter(out).emitJavadoc(javadocCodeBlock);
-        assertThat(out.toString()).isEqualTo(
+        assertThat(out).hasToString(
                 "/**\n" +
                         " * A\n" +
                         " *\n" +
