@@ -16,6 +16,9 @@
 package com.squareup.javapoet;
 
 import com.google.common.io.ByteStreams;
+import org.junit.jupiter.api.Test;
+import org.junit.rules.TemporaryFolder;
+
 import java.io.IOException;
 import java.net.URI;
 import java.util.Collections;
@@ -30,93 +33,90 @@ import javax.tools.JavaFileObject.Kind;
 import javax.tools.StandardJavaFileManager;
 import javax.tools.StandardLocation;
 import javax.tools.ToolProvider;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
 
-import static com.google.common.truth.Truth.assertThat;
 import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.assertj.core.api.Assertions.assertThat;
 
-@RunWith(JUnit4.class)
-public class FileReadingTest {
-  
-  // Used for storing compilation output.
-  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+public class FileReadingTest extends AbstractTemporaryFolder {
 
-  @Test public void javaFileObjectUri() {
-    TypeSpec type = TypeSpec.classBuilder("Test").build();
-    assertThat(JavaFile.builder("", type).build().toJavaFileObject().toUri())
-        .isEqualTo(URI.create("Test.java"));
-    assertThat(JavaFile.builder("foo", type).build().toJavaFileObject().toUri())
-        .isEqualTo(URI.create("foo/Test.java"));
-    assertThat(JavaFile.builder("com.example", type).build().toJavaFileObject().toUri())
-        .isEqualTo(URI.create("com/example/Test.java"));
-  }
-  
-  @Test public void javaFileObjectKind() {
-    JavaFile javaFile = JavaFile.builder("", TypeSpec.classBuilder("Test").build()).build();
-    assertThat(javaFile.toJavaFileObject().getKind()).isEqualTo(Kind.SOURCE);
-  }
-  
-  @Test public void javaFileObjectCharacterContent() throws IOException {
-    TypeSpec type = TypeSpec.classBuilder("Test")
-        .addJavadoc("Pi\u00f1ata\u00a1")
-        .addMethod(MethodSpec.methodBuilder("fooBar").build())
-        .build();
-    JavaFile javaFile = JavaFile.builder("foo", type).build();
-    JavaFileObject javaFileObject = javaFile.toJavaFileObject();
-    
-    // We can never have encoding issues (everything is in process)
-    assertThat(javaFileObject.getCharContent(true)).isEqualTo(javaFile.toString());
-    assertThat(javaFileObject.getCharContent(false)).isEqualTo(javaFile.toString());
-  }
-  
-  @Test public void javaFileObjectInputStreamIsUtf8() throws IOException {
-    JavaFile javaFile = JavaFile.builder("foo", TypeSpec.classBuilder("Test").build())
-        .addFileComment("Pi\u00f1ata\u00a1")
-        .build();
-    byte[] bytes = ByteStreams.toByteArray(javaFile.toJavaFileObject().openInputStream());
-    
-    // JavaPoet always uses UTF-8.
-    assertThat(bytes).isEqualTo(javaFile.toString().getBytes(UTF_8));
-  }
-  
-  @Test public void compileJavaFile() throws Exception {
-    final String value = "Hello World!";
-    TypeSpec type = TypeSpec.classBuilder("Test")
-        .addModifiers(Modifier.PUBLIC)
-        .addSuperinterface(ParameterizedTypeName.get(Callable.class, String.class))
-        .addMethod(MethodSpec.methodBuilder("call")
-            .returns(String.class)
-            .addModifiers(Modifier.PUBLIC)
-            .addStatement("return $S", value)
-            .build())
-        .build();
-    JavaFile javaFile = JavaFile.builder("foo", type).build();
 
-    JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
-    DiagnosticCollector<JavaFileObject> diagnosticCollector = new DiagnosticCollector<>();
-    StandardJavaFileManager fileManager = compiler.getStandardFileManager(diagnosticCollector, 
-        Locale.getDefault(), UTF_8);
-    fileManager.setLocation(StandardLocation.CLASS_OUTPUT,
-        Collections.singleton(temporaryFolder.newFolder()));
-    CompilationTask task = compiler.getTask(null, 
-        fileManager,
-        diagnosticCollector,
-        Collections.emptySet(),
-        Collections.emptySet(),
-        Collections.singleton(javaFile.toJavaFileObject()));
-    
-    assertThat(task.call()).isTrue();
-    assertThat(diagnosticCollector.getDiagnostics()).isEmpty();
+    @Test
+    public void javaFileObjectUri() {
+        TypeSpec type = TypeSpec.classBuilder("Test").build();
+        assertThat(JavaFile.builder("", type).build().toJavaFileObject().toUri())
+                .isEqualTo(URI.create("Test.java"));
+        assertThat(JavaFile.builder("foo", type).build().toJavaFileObject().toUri())
+                .isEqualTo(URI.create("foo/Test.java"));
+        assertThat(JavaFile.builder("com.example", type).build().toJavaFileObject().toUri())
+                .isEqualTo(URI.create("com/example/Test.java"));
+    }
 
-    ClassLoader loader = fileManager.getClassLoader(StandardLocation.CLASS_OUTPUT);
-    Callable<?> test = Class.forName("foo.Test", true, loader)
-            .asSubclass(Callable.class)
-            .getDeclaredConstructor()
-            .newInstance();
-    assertThat(Callable.class.getMethod("call").invoke(test)).isEqualTo(value);
-  }
+    @Test
+    public void javaFileObjectKind() {
+        JavaFile javaFile = JavaFile.builder("", TypeSpec.classBuilder("Test").build()).build();
+        assertThat(javaFile.toJavaFileObject().getKind()).isEqualTo(Kind.SOURCE);
+    }
+
+    @Test
+    public void javaFileObjectCharacterContent() throws IOException {
+        TypeSpec type = TypeSpec.classBuilder("Test")
+                .addJavadoc("Pi\u00f1ata\u00a1")
+                .addMethod(MethodSpec.methodBuilder("fooBar").build())
+                .build();
+        JavaFile javaFile = JavaFile.builder("foo", type).build();
+        JavaFileObject javaFileObject = javaFile.toJavaFileObject();
+
+        // We can never have encoding issues (everything is in process)
+        assertThat(javaFileObject.getCharContent(true)).isEqualTo(javaFile.toString());
+        assertThat(javaFileObject.getCharContent(false)).isEqualTo(javaFile.toString());
+    }
+
+    @Test
+    public void javaFileObjectInputStreamIsUtf8() throws IOException {
+        JavaFile javaFile = JavaFile.builder("foo", TypeSpec.classBuilder("Test").build())
+                .addFileComment("Pi\u00f1ata\u00a1")
+                .build();
+        byte[] bytes = ByteStreams.toByteArray(javaFile.toJavaFileObject().openInputStream());
+
+        // JavaPoet always uses UTF-8.
+        assertThat(bytes).isEqualTo(javaFile.toString().getBytes(UTF_8));
+    }
+
+    @Test
+    public void compileJavaFile() throws Exception {
+        final String value = "Hello World!";
+        TypeSpec type = TypeSpec.classBuilder("Test")
+                .addModifiers(Modifier.PUBLIC)
+                .addSuperinterface(ParameterizedTypeName.get(Callable.class, String.class))
+                .addMethod(MethodSpec.methodBuilder("call")
+                        .returns(String.class)
+                        .addModifiers(Modifier.PUBLIC)
+                        .addStatement("return $S", value)
+                        .build())
+                .build();
+        JavaFile javaFile = JavaFile.builder("foo", type).build();
+
+        JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();
+        DiagnosticCollector<JavaFileObject> diagnosticCollector = new DiagnosticCollector<>();
+        StandardJavaFileManager fileManager = compiler.getStandardFileManager(diagnosticCollector,
+                Locale.getDefault(), UTF_8);
+        fileManager.setLocation(StandardLocation.CLASS_OUTPUT,
+                Collections.singleton(newFolder()));
+        CompilationTask task = compiler.getTask(null,
+                fileManager,
+                diagnosticCollector,
+                Collections.emptySet(),
+                Collections.emptySet(),
+                Collections.singleton(javaFile.toJavaFileObject()));
+
+        assertThat(task.call()).isTrue();
+        assertThat(diagnosticCollector.getDiagnostics()).isEmpty();
+
+        ClassLoader loader = fileManager.getClassLoader(StandardLocation.CLASS_OUTPUT);
+        Callable<?> test = Class.forName("foo.Test", true, loader)
+                .asSubclass(Callable.class)
+                .getDeclaredConstructor()
+                .newInstance();
+        assertThat(Callable.class.getMethod("call").invoke(test)).isEqualTo(value);
+    }
 }
